Move PlayerConnection and PlayerManager to Shared
